rename: README.md에 정리한 폴더 구조 반영
