+++ src/components/Main/InfluencerSection/InfluencerItem.tsx
