+++ src/components/Main/SpotSection/index.tsx
