+++ src/pages/Main/index.tsx
@@ -1,8 +1,8 @@
 import styled from 'styled-components';
 
-import BaseLayout from '@/components/features/Main/BaseLayout';
-import InfluencerSection from '@/components/features/Main/InfluencerSection';
-import MainBanner from '@/components/features/Main/MainBanner';
+import BaseLayout from '@/components/Main/BaseLayout';
+import InfluencerSection from '@/components/Main/InfluencerSection';
+import MainBanner from '@/components/Main/MainBanner';
 
 import { BannerData, InfluencerData } from '@/types';
